requirements.txt
streamlit==1.31.0
langchain==0.3.0
langchain-openai==0.2.0
requests==2.31.0
auth0-python==4.7.0